# [Core] performance improvements around bindings