--- a/Xamarin.Forms.Core/BindingBase.cs
+++ b/Xamarin.Forms.Core/BindingBase.cs
@@ -87,6 +87,7 @@ namespace Xamarin.Forms
 			SynchronizedCollections.Add(collection, new CollectionSynchronizationContext(context, callback));
 		}
 
+		[MethodImpl(MethodImplOptions.AggressiveInlining)]
 		protected void ThrowIfApplied()
 		{
 			if (IsApplied)