--- a/Xamarin.Forms.Core/BindingBaseExtensions.cs
+++ b/Xamarin.Forms.Core/BindingBaseExtensions.cs
@@ -1,7 +1,10 @@
-﻿namespace Xamarin.Forms
+﻿using System.Runtime.CompilerServices;
+
+namespace Xamarin.Forms
 {
 	static class BindingBaseExtensions
 	{
+		[MethodImpl(MethodImplOptions.AggressiveInlining)]
 		public static BindingMode GetRealizedMode(this BindingBase self, BindableProperty property)
 		{
 			return self.Mode != BindingMode.Default ? self.Mode : property.DefaultBindingMode;