@jonathanpeppers 2020-01-07
> If it's under 0.5% then I'd suggest weighing the improvement against the possibility of regression. 

Can you elaborate what this means? Run the numbers _again_?

I wrote both a set of benchmarks and a sample app here, both show improvements.