@StephaneDelcroix 2020-01-08
it'd be good to time this vs compiled bindings too